Part 5 - Flow Control
It was suggested to use: OneOf, FluentResults, ErrorOr packages to return errors in functional programming maner but I believe it is an overkill. It probably gotta be such a pain in the ass to write code with all these error statements. But if all infrustructure is already defined and you just need to use this method for your concrete functionality... Perhaps it may not be that bad after all.